+++ CloneCoding/carrot/src/main/java/com/sopt/carrotmarket/shared/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package com.sopt.carrotmarket.domain;
+package com.sopt.carrotmarket.shared;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
